Moved a file './pgp/armor.ts' to ./pgp-armor.ts
